requirements.txt
streamlit
openai
PyPDF2
python-dotenv
pinecone-client